Finish code in unk_0207F42C.s

+++ include/field_move.h
@@ -4,7 +4,7 @@
 #include "script.h"
 #include "party_menu.h"
 
-void *PartyMenu_GetFieldMoveFunc(int funcType, int fieldMoveIdx);
+void *PartyMenu_GetFieldMoveFunc(u16 funcType, u16 fieldMoveIdx);
 void FieldSystem_MakeFieldMoveCheckData(FieldSystem *fieldSystem, struct FieldMoveCheckData *checkData);
 
 #endif //POKEHEARTGOLD_FIELD_MOVE_H
